+++ ibctesting/coordinator.go
@@ -9,26 +9,8 @@ import (
 
 var globalStartTime = time.Date(2020, 1, 2, 0, 0, 0, 0, time.UTC)
 
-// NewCoordinator initializes Coordinator with N EVM TestChain's (Evmos apps)
-func NewEVMCoordinator(t *testing.T, n int) *ibctesting.Coordinator {
-	chains := make(map[string]*ibctesting.TestChain)
-	coord := &ibctesting.Coordinator{
-		T:           t,
-		CurrentTime: globalStartTime,
-	}
-
-	ibctesting.DefaultTestingAppInit = DefaultTestingAppInit
-	for i := 1; i <= n; i++ {
-		chainID := GetChainID(i)
-		chains[chainID] = NewTestChain(t, coord, chainID)
-	}
-	coord.Chains = chains
-
-	return coord
-}
-
-// NewMixedCoordinator initializes Coordinator with N EVM TestChain's (Evmos apps) and M Cosmos chains (Simulation Apps)
-func NewMixedCoordinator(t *testing.T, nEVMChains, mCosmosChains int) *ibctesting.Coordinator {
+// NewCoordinator initializes Coordinator with N EVM TestChain's (Evmos apps) and M Cosmos chains (Simulation Apps)
+func NewCoordinator(t *testing.T, nEVMChains, mCosmosChains int) *ibctesting.Coordinator {
 	chains := make(map[string]*ibctesting.TestChain)
 	coord := &ibctesting.Coordinator{
 		T:           t,
@@ -39,7 +21,7 @@ func NewMixedCoordinator(t *testing.T, nEVMChains, mCosmosChains int) *ibctestin
 	ibctesting.DefaultTestingAppInit = DefaultTestingAppInit
 
 	for i := 1; i <= nEVMChains; i++ {
-		chainID := GetChainID(i)
+		chainID := ibctesting.GetChainID(i)
 		chains[chainID] = NewTestChain(t, coord, chainID)
 	}