+++ x/erc20/keeper/evm_test.go
@@ -23,7 +23,7 @@ func (suite *KeeperTestSuite) TestQueryERC20() {
 		},
 		{
 			"ok",
-			func() { contract = suite.DeployContract("coin", "token", erc20Decimals) },
+			func() { contract, _ = suite.DeployContract("coin", "token", erc20Decimals) },
 			true,
 		},
 	}
@@ -66,7 +66,8 @@ func (suite *KeeperTestSuite) TestCallEVM() {
 		suite.SetupTest() // reset
 
 		erc20 := contracts.ERC20MinterBurnerDecimalsContract.ABI
-		contract := suite.DeployContract("coin", "token", erc20Decimals)
+		contract, err := suite.DeployContract("coin", "token", erc20Decimals)
+		suite.Require().NoError(err)
 		account := tests.GenerateAddress()
 
 		res, err := suite.app.Erc20Keeper.CallEVM(suite.ctx, erc20, types.ModuleAddress, contract, tc.method, account)