+++ x/incentives/keeper/integration_test.go
@@ -12,6 +12,37 @@ import (
 	types "github.com/tharsis/evmos/v2/x/incentives/types"
 )
 
+var _ = Describe("Performing EVM transactions", Ordered, func() {
+
+	BeforeEach(func() {
+		s.SetupTest()
+
+		params := s.app.Erc20Keeper.GetParams(s.ctx)
+		params.EnableEVMHook = false
+		s.app.Erc20Keeper.SetParams(s.ctx, params)
+	})
+
+	// Epoch mechanism for triggering allocation and distribution
+	Context("with the ERC20 module disabled", func() {
+		It("should be successful", func() {
+			_, err := s.DeployContract("coin", "token", erc20Decimals)
+			Expect(err).To(BeNil())
+		})
+	})
+
+	Context("with the ERC20 module enabled", func() {
+		BeforeEach(func() {
+			params := s.app.Erc20Keeper.GetParams(s.ctx)
+			params.EnableEVMHook = true
+			s.app.Erc20Keeper.SetParams(s.ctx, params)
+		})
+		It("should be successful", func() {
+			_, err := s.DeployContract("coin", "token", erc20Decimals)
+			Expect(err).To(BeNil())
+		})
+	})
+})
+
 var _ = Describe("Distribution", Ordered, func() {
 	var (
 		balanceBefore  sdk.Coin
@@ -32,10 +63,12 @@ var _ = Describe("Distribution", Ordered, func() {
 		participantAcc = sdk.AccAddress(s.address.Bytes())
 
 		// Deploy contract
-		contractAddr = s.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
+		var err error
+		contractAddr, err = s.DeployContract(erc20Name, erc20Symbol, erc20Decimals)
+		s.Require().NoError(err)
 
 		// Create incentive
-		_, err := s.app.IncentivesKeeper.RegisterIncentive(
+		_, err = s.app.IncentivesKeeper.RegisterIncentive(
 			s.ctx,
 			contractAddr,
 			mintAllocations,