+++ x/withdraw/types/keys.go
@@ -11,3 +11,13 @@ const (
 	// RouterKey to be used for message routing
 	RouterKey = ModuleName
 )
+
+// prefix bytes for the withdraw module persistent store
+const (
+	prefixBech32HRP = iota + 1
+)
+
+// KVStore key prefixes
+var (
+	KeyPrefixBech32HRP = []byte{prefixBech32HRP}
+)